chore(next-auth): move next auth to app router
