+++ src/app/api/auth/[...nextauth]/route.ts
@@ -84,4 +84,6 @@ export const authOptions: AuthOptions = {
   },
 };
 
-export default NextAuth(authOptions);
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
