+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import TrpcProvider from "@/components/context/TrpcProvider";
 import NextAuthProvider from "@/components/context/NextAuthProvider";
 import NavbarApp from "@/components/common/Navbar/NavbarApp";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 const roboto = Roboto({
   variable: "--font-roboto",
