+++ src/server/actions/submitApplication.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { prisma } from "@/services/prisma";
 import isApplicationComplete from "@/server/services/helpers/isApplicationComplete";
 import { revalidatePath } from "next/cache";
