+++ src/server/getters/applicationFormStep.ts
@@ -1,6 +1,6 @@
 import { prisma } from "@/services/prisma";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { FormFieldType, FormFieldTypeEnum } from "@/services/types/formFields";
 
 export type FormFieldValueType = string | boolean | null;
