+++ src/services/helpers/localData/updateLocalApplicationData.ts
@@ -1,10 +1,12 @@
 import { SaveApplicationStepFormInput } from "@/server/actions/saveApplicationStepForm";
 import getLocalApplicationData from "@/services/helpers/localData/getLocalApplicationData";
 
-const updateLocalApplicationData = (input: SaveApplicationStepFormInput) => {
+const updateLocalApplicationData = ({
+  fieldValues,
+}: SaveApplicationStepFormInput) => {
   const localApplicationData = getLocalApplicationData();
   const newData = localApplicationData ?? [];
-  for (const field of input) {
+  for (const field of fieldValues) {
     const index = newData.findIndex((f) => f.fieldId === field.fieldId);
     if (index === -1) {
       newData.push(field);
