+++ src/services/helpers/requireAuthServerSideProps.ts
@@ -1,6 +1,6 @@
 import { GetServerSidePropsContext } from "next";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 const requireAuthServerSideProps = async (
   context: GetServerSidePropsContext
