+++ src/services/helpers/requireOrganizerApp.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { prisma } from "@/services/prisma";
 
 const requireOrganizerApp = async (): Promise<boolean> => {
