+++ src/services/helpers/requireOrganizerServerSideProps.ts
@@ -1,6 +1,6 @@
 import { GetServerSidePropsContext } from "next";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { prisma } from "@/services/prisma";
 
 const requireOrganizerServerSideProps = async (
