+++ tsconfig.json
@@ -40,6 +40,7 @@
     ".next/types/**/*.ts"
   ],
   "exclude": [
-    "node_modules"
+    "node_modules",
+    ".next/types/app/api/auth/**/*.ts"
   ]
 }
